Improve reading rate in VideoReader
Instead of reading one byte at once, multiple thousands of bytes are
read at once now. This is an amount of bytes read which most machines
should be capable of.
The read data is stored in a buffer, which content is searched for the
JPEG markers SOI and EOI. Dependent on which information found a JPEG
image (or a part of it) is stored in an image buffer. If there is data
left in the buffer after the end of the image, the remaining data is
shifted to the start of the buffer. The next time when bytes are read
from the file, the read data is stored with an offset so that the
shifted data is not overwritten.
This continues until EOF is reached.