Update statistics in Client GUI more seldom
The statistics string in the Client's GUI was updated everytime a new
frame was shown. This was approximatly 25 times per second and resulted
in unreasonable CPU load.
Now the statistics are updated every 5th frame, which still is quite
fast. But it is readability increased and the CPU load sank.