Introduce functionallity for parsing AVI metadata
To parse the AVI metadata the AviMetadataParser is created. It checks
for the RIFF file header, the main AVI header and a stream header.
Latter is completely read and parsed. The framerate and duration of the
video is stored in a VideoMetadata object. The VideoMetadata class
should be used for all cases where metadata is extracted from files and
handled in the program.