Enable compatibility with VLC Media Player
In the current usecase each video frame is send as one RTP packet. RFC
2435 states that every last packet belonging to a single frame has to
set the marker in the RTP packet header. That's why in our case every
RTP packet has to set the marker bit.

A second change is done in the timestamp of RTP packets. The previous
mentioned RFC defines also that the RTP timestamp of JPEG packets has to
be in units of 90000Hz. This number results in a whole number for
framerates of 25 or 30 Hz. While this information has previously
included in the SDP description, the change was not applied to the
actual timestamps. Now the timestamp is calculated as:
    ts = <frame number> * 90000Hz / <framerate>Hz

Both changed together (and also some previous changes) result in the
compatibility with the VLC Media Player.