Move client side RTP functionality to RtpHandler
The client side RTP handling (registration of received packets,
provision of images for displaying) is moved to the RtpHandler.
Now the client has a simple call to the RtpHandler for processing a
received packet (including registration for error correction) and
another simple call for getting an image for displaying in the GUI.

A lot of functionality is removed from the FecHandler, which previously
handled most of the RTP processing. Now the FecHandler just provides FEC
error correction.

Besides the functionality for RTP processing the statistics are also
removed from the FecHandler. Now there is a simple class for handling
such statistic values which makes it easier for the client to get those
from the RtpHandler.