Connect Radiobuttons with RtpHandler functionality
The Radiobuttons in Server and Client are now connected with the
encryption functionality in the RtpHandler.
Latter now checks for the encryption mode when a packet is created (RTP
or FEC) or processed (a received packet). If the encryption mode is
supported, a special handler is called for encrypting or decrypting the
packet. Currently the options are no encryption or SRTP encryption. JPEG
encryption is present as placeholder and will be included in future.

To have the RtpHandler present when initializing the GUI and thus
initializing the encryption mode, Server and Client had to initialize
their RtpHandler in the constructor. An additional method is added to
provide the same possibilities for enabling or disabling the FEC error
correction in the Client outside it's constructor.

The RtpHandler also provides default values for a master key and master
salt, to avoid the need of session initialization and key management.
This is open to add by future work.