Implement replacement attack for JPEG encryption
Because the quantization tables were encrypted and the JPEG standard
specifies a set of example tables which achieve good results a
replacement attack can be executed.
The aim of the replacement attack is to replace the encrypted
quantization tables with these from the JPEG standard. Although the
tables from the standard may not be exactly the original ones, they may
give the opportunity to display an image very similar to the original
one.

This patch also includes a test for the replacement attack.

Additionally the Client and RtpHandler are extended to support this
attack at a RTP stream.